Added float suupport for the weight column

+++ models/order.py
@@ -10,7 +10,7 @@ class purchase_order_line(models.Model):
 
     leadtime = fields.Integer('Leadtime', default=10, help='Requested leadtime in days.')
     units_shipped = fields.Integer('Units Shipped', help='Units Shipped',default=0)
-    weight = fields.Integer('Weight', help='Weight',default=0)
+    weight = fields.Float('Weight', help='Weight',default=0)
     collies = fields.Integer('# Collies', help='# Collies',default=0)
     units_in_stock = fields.Integer('Units in Stock', help='Units in Stock',default=0)
     batch_number = fields.Char('Batch Number', help='Batch Number',size=24,default='N/A')
